Rename `src/annotator/util/{dom -> dom-element.js}`
Rename this module to distinguish it from `src/sidebar/util/dom.js`
which are generic DOM utilities that are not ponyfills for standard
methods.

See https://github.com/hypothesis/client/pull/2106/files#r420296891
