+++ src/annotator/util/dom-element.js
@@ -1,4 +1,10 @@
 /**
+ * Ponyfills [1] for DOM element methods that are not supported in all browsers.
+ *
+ * [1] https://github.com/sindresorhus/ponyfill
+ */
+
+/**
  * Implementation of `element.closest(selector)`. This is used to support browsers
  * (IE 11) that don't have a native implementation.
  */
